+++ src/lib.rs
@@ -19,7 +19,7 @@ pub use client::Client;
 pub use err::AppErr;
 pub use fluff::Fluff;
 pub use gem_bytes::GemBytes;
-pub use gem_call::GemCall;
+pub use gem_call::{GemCall, Static};
 
 /// Procedural macro that must be used in implementations of [`GemBytes`].
 pub use async_trait::async_trait;